carry-server/requirements.txt
alabaster==0.7.12
aniso8601==8.0.0
astroid==2.4.0
attrs==19.3.0
autopep8==1.5.2
Babel==2.8.0
bcrypt==3.1.7
cachetools==4.1.0
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
click==7.1.2
commonmark==0.9.1
cryptography==2.9.2
dnspython==1.16.0
docutils==0.16
dominate==2.5.1
email-validator==1.0.5
Flask==1.1.2
Flask-Admin==1.5.6
Flask-Bcrypt==0.7.1
flask-googlemaps==0.4.0
Flask-Login==0.5.0
flask-marshmallow==0.12.0
Flask-SocketIO==4.3.0
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.3
futures==3.1.1
google-api-core==1.19.0
google-api-python-client==1.9.1
google-assistant-grpc==0.2.1
google-assistant-sdk==0.6.0
google-auth==1.16.1
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.1
googleapis-common-protos==1.52.0
greenlet==0.4.15
grpcio==1.29.0
httplib2==0.18.1
idna==2.9
imagesize==1.2.0
importlib-metadata==1.6.0
is-disposable-email==1.0.0
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
marshmallow==3.5.2
marshmallow-sqlalchemy==0.23.0
mccabe==0.6.1
more-itertools==8.2.0
oauthlib==3.1.0
packaging==20.3
pathlib2==2.3.5
pkg-resources==0.0.0
pluggy==0.13.1
protobuf==3.12.2
py==1.8.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAudio==0.2.11
pycodestyle==2.5.0
pycparser==2.20
pycrypto==2.6.1
Pygments==2.6.1
pylint==2.5.0
PyMySQL==0.9.3
pyOpenSSL==19.1.0
pyparsing==2.4.7
pypng==0.0.20
PyQRCode==1.2.1
python-engineio==3.12.1
python-socketio==4.5.1
pytz==2020.1
recommonmark==0.6.0
requests==2.23.0
requests-oauthlib==1.3.0
rinoh-typeface-dejavuserif==0.1.1
rinoh-typeface-texgyrecursor==0.1.1
rinoh-typeface-texgyreheros==0.1.1
rinoh-typeface-texgyrepagella==0.1.1
rinohtype==0.4.0
rsa==4.0
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
six==1.15.0
snowballstemmer==2.0.0
sounddevice==0.3.15
SpeechRecognition==3.8.1
Sphinx==3.0.3
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
SQLAlchemy==1.3.16
sqlalchemy-filters==0.11.0
tenacity==4.12.0
toml==0.10.0
typed-ast==1.4.1
uritemplate==3.0.1
urllib3==1.25.9
visitor==0.1.3
wcwidth==0.1.9
Werkzeug==1.0.1
wrapt==1.12.1
WTForms==2.3.1
wtforms-validators==1.0.0
zipp==3.1.0
